moving all methods relating to halo orbits non-specific to WFIRST from WFIRSTObservatoryL2 to new class ObservatoryL2Halo
